Implement Authorization using Policy

+++ app/Http/Controllers/QuestionsController.php
@@ -7,6 +7,15 @@ use App\Http\Requests\AskQuestionRequest;
 
 class QuestionsController extends Controller
 {
+
+    /**
+     * QuestionsController constructor.
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => 'index', 'show']);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -61,9 +70,12 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        /*
         if (\Gate::denies('update-question', $question)) {
             abort(403, "Access Denied");
         }
+        */
+        $this->authorize('update', $question);
         return view('questions.edit', ['question' => $question]);
     }
 
@@ -76,9 +88,13 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        /*
         if (\Gate::denies('update-question', $question)) {
             abort(403, "Access Denied");
         }
+        */
+        $this->authorize('update', $question);
+
         $question->update($request->only('title', 'body'));
         return redirect()->route('questions.index')->with('success', 'Your question has been updated');
     }
@@ -91,9 +107,13 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        /*
         if (\Gate::denies('delete-question', $question)) {
             abort(403, "Access Denied");
         }
+        */
+        $this->authorize('delete', $question);
+
         $question->delete();
         return redirect()->route('questions.index')->with('success', 'Your question has been deleted.');
     }
